pyproject.toml
[project]
name = "SRBenchmark"
version = "1.0.0"
description = "Automated benchmark for Super-Resolution neural network models"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "einops>=0.8.2",
    "matplotlib>=3.10.8",
    "numpy>=2.4.4",
    "omegaconf>=2.3.0",
    "opencv-python>=4.13.0.92",
    "packaging>=26.1",
    "pandas>=3.0.2",
    "ptflops>=0.7.5",
    "pyiqa>=0.1.15.post2",
    "pyyaml>=6.0.3",
    "safetensors>=0.7.0",
    "torch>=2.11.0",
    "torchvision>=0.26.0",
    "tqdm>=4.67.3",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-mock>=3.15.1",
]

[tool.ruff]
line-length = 120

